fix(rosetta): stop skipping example values (#3128)
In CDK, we use the `@example` tag a lot to indicate example values
of some property, instead of code examples. For example, a `bucketArn`
property might have as annotation `@example arn:aws:s3:....`.

Rosetta would try to compile those, and fail. To get around this, we would skip
values that *looked* like they weren't source in `rosetta extract`.

This doesn't combine well with pacmak live-translation. We are either forced to
try and compile them again during pacmak (adding compile time), or fail
pacmak as soon as we encounter these. And all of this is to work around
abuse of the `@example` tag for something it's not intended for.

So stop treating these examples specially. They will just fail compilation
unless rewritten. This will force us to deal with them properly in the
CDK source code.



---

By submitting this pull request, I confirm that my contribution is made under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0