fix(jsii): incorrectly allowed unexported type in constructor params
The compiler incorrectly allowed un-epxorted types to be used in a
constructor parameter, resulting in invalid assemblies (where a
reference to the un-exported type exists, but the type itself does not).

Added a test that validates this is now checked against, and improved
the error experience by providing source anchors for where the
problematic type is being used.