fix(rosetta): `didCompile` evaluates to true when compilation not attempted (#3149)
When `rosetta:extract` is run, the property `didCompile` evalutes to `true`
even when compilation was not attempted. This causes lots of confusion.
The correct way to handle this is to return `undefined` when compilation is
not attempted. Then, (I assume) `didCompile` will not show up in the tablet
file at all when compilation is not asked for (via `--compile` or `strict`.

---

By submitting this pull request, I confirm that my contribution is made under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0