chore(rosetta): suppress diagnostics after a certain point
In a regular build, Rosetta might throw up hundreds or thousands
of warnings about failed translations. Past a certain point, there's
just no use to so many error messages any more.

Limit the output to 50 at a time.

**"But Rico, why no flag to show them all?"** Well, either a human is
going to actually look at them and fix them, in which case they're
not going to do more than a chunk at a time and re-run Rosetta anyway...
or nobody is going to look at them and then what's the point of
being able to show all? The count is good enough to get a sense for
how bad things are.