fix(rosetta): types from submodules not recognized properly (#3174)
Rosetta used to store a guessed version of the type fqn when running `rosetta extract`. These guessed fqns are correct for v1 but break down on v2, since they do not properly account for namespaces. This PR correctly determines the fqn of a type by computing the symbolId, loading the relevant assembly, and matching the symbolId with the actual fqn.

---

By submitting this pull request, I confirm that my contribution is made under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0