fix: C# NamespaceDoc emitted to wrong location
The namespace determination for the `NamespaceDoc` class generated for
submodules was incorrect, as the `jsiiNs` passed to the type resovler
was absolute, instead of being the required assembly-relative form.

Additionally, removed the `readme` and `locationInSource` property from
submodule descriptors included in the jsii assembly under
`dependencyClosure`, as those are not necessary and needlessly bloat the
assembly document.