feat(rosetta):  metadata tag for fixtures in docs (#3200)
Currently the best thing to do to specify metadata in `@example` docs is to supply it in the example like so:

```
/**
* @example
* /// metadata
* my example
*/
```

This works, but is cumbersome because the `/// metadata` line gets picked up in typescript doc gen and gets displayed as part of the example in docs. Because of backwards compatibility, this functionality stays, but this PR introduces a better way to supply example metadata via the `@exampleMetadata` tag. For example:

```
/**
 * @exampleMetadata nofixture
 * @example 
 * my example
 */
```

`@exampleMetadata` will be read by rosetta and the necessary fixtures supplied. While this PR focuses on adding functionality for fixture metadata, it theoretically should work on any kind of metadata we may want to add in the future.

---

By submitting this pull request, I confirm that my contribution is made under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0