fix(rosetta): use `--compile` flag by default
The way we are trending, there is no good reason anymore to NOT compile
your examples. We have good support for parallelism and caching now,
so the argument that it takes too much time to do the additional
analysis no longer really holds.

Also, why should people have to remember to pass this flag for default
that should arguably be the default?

Make it the default.