fix(pacmak): don't automatically translate examples without asking
Given the current state of example tooling, the existing default
was a bad choice.

It forces everyone who is publishing jsii packages everywhere to
spend CPU cycles to do example parsing and translation which is
bound to fail anyway (as the fixtures and compilation directory
will probably not be set up for success, and translation
must proceed in a single-threaded fashion).

The new default is to use translations if they exist, and otherwise
don't do translations.

* Builds that care (i.e., mostly the CDK build), will be using `rosetta
  extract` already to be efficient and correct about compilation, and
  the tablet files will be picked up as usual.
* Builds that don't care will get more efficient.

BREAKING CHANGES: Pacmak no longer tries to translate examples
by default: this feature is now opt-in.