fix(rosetta): classes are not correctly identified if package uses an outDir (#3225)
`rosetta extract` breaks down when a project structure includes `outDir` or `rootDir` in the `tsconfig` file. Given such a project structure, `extract` may find symbolIds that look like `outDir/index:ClassA` or `rootDir/index:ClassA`. Since the jsii assembly is created from the source files in the `rootDir`, `rootDir/index:ClassA` is the symbolId we want. So if the `outDir` is specified, we swap it with the specified `rootDir` (and if `rootDir` is not specified, we guess at it the same way node does). This ensures that `rosetta extract` is structure agnostic.

Co-authored-by: Rico Huijbers <rix0rrr@gmail.com>