fix(jsii): handle imports from libraries compiled with old jsii
The fix introduced in #3233 relied on the jsii libraries being recompiled with
symbolid information. This information was introduced in jsii 1.39.0.

With the code from #3233, symbols from libraries compiled with older jsii
versions could not be located and the compilation would fail.

In this PR, restore the compatibility by falling back to the same FQN
that the original code would have guessed. It will not be correct
if the type comes from a submodule, but only CDK is using submodules
and CDK is using a new jsii compiler.