fix(rosetta): `rosetta extract` fails if run after `rosetta infuse` (#3248)
There is a valid use case to run `rosetta extract` after `rosetta infuse`.
However, we run into problems because we `fixturize` snippets
before we check the cache, and sometimes `fixturizing` will fail for
infused snippets (specifically, literate sources like `integ.blah.lit.ts`
will fail).

We do not actually need to successfully `fixturize` infused examples.
We rely on these examples to be in the cached tablet file and thus
do not need to be compiled and translated again. So we introduce
a new metadata tag, `infuse=true` for all infused examples. This tag, 
when consumed by `extract`, will skip any errors associated with
that snippet. 

ALSO: increased test timeout from `30000` to `50000` to allow enough
time for mac PR test to succeed.

---

By submitting this pull request, I confirm that my contribution is made under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0