chore: tweak auto-approval workflow
The automated approval workflow will no longer emit an approval if the
PR already has reviews. In case the PR was issued by @dependabot, it
will approve with `@dependabot squash and merge` as the review body,
which requests @dependabot to automatically squash and merge the PR once
all CI checks successfully complete.

This should help remove the need for manual intervention on most
dependabot PRs.