fix(pacmak): fails on bundled dependency without entry point
`projen` uses a bundled dependency called `shx`, and this library does
not have an entry point. It exposes some command-line tools and nothing
else.

This conflicts with our dependency finding mechanism, which performs
a `require()` on the library to find it.

Ignore bundled dependencies in `pacmak`, it never needs to pack them
individually anyway.