fix(rosetta): diagnostics from infused snippets were not ignored (#3282)
Recently we implemented the `infused` metadata that turns `strict=false`
and `loose=true` for the specific snippet. However we still output the
diagnostics along with other snippets (with `isError=true`). One thing we
could do is make these diagnostics `isError=false`, but I argue that we
want to ignore these diagnostics entirely. They aren't actually diagnostics;
we will find them in the cache and they do compile.

---

By submitting this pull request, I confirm that my contribution is made under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0