chore: auto approve workflow broken (#3288)
The auto-approve workflow is currently [failing](https://github.com/aws/jsii/runs/4618073266?check_suite_focus=true) with:

```console
Error: Input required and not supplied: github-token
Error: Input required and not supplied: github-token
    at Object.getInput (/home/runner/work/_actions/actions/github-script/0.3.0/node_modules/@actions/core/lib/core.js:73:15)
    at main (/home/runner/work/_actions/actions/github-s
```

This because version `0.3.0` of the action requires a `github-token`, and we don't always provide it. 

Instead of adding it, I upgraded to version `5.0.0` of the action, which no longer requires it.
Some adjustments to the implementation were also necessary due to [breaking changes](https://github.com/actions/github-script#breaking-changes-in-v5). 

---

By submitting this pull request, I confirm that my contribution is made under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0