fix(rosetta): infused snippets not returned from cache (#3291)
Infused snippets do not have the right full source since we are adding the local fixture to the snippet
rather than the fixture from where the snippet came from. Since there is no hope in actually translating
an infused snippet (unless it is in the same directory), we might as well take any cache result as is.
So this PR adds logic that returns the cache result for infused snippets always.

This was tested on my local machine and dropped all errors from infused snippets. There are still errors
from directories that are not yet strict, which is understandable.

Also, the infuse command was missing `cache-from` argument that is present in `infuseOptions`.
We were using those options as part of `rosetta extract --infuse` but should be available via `rosetta infuse`
as well.

Also fixed up minor mistake in tests where I was testing for `infused=true` instead of `infused=''`.

---

By submitting this pull request, I confirm that my contribution is made under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0