fix(jsii): cannot use submodule types from assemblies before 1.39.0 (#3306)
When trying to use submodule types from assemblies compiled with
a jsii version `< 1.39.0`, you will see the following error:

```
Unable to resolve type "<module>.<type>". It may be @internal or not
exported from the module's entry point (as configured in "package.json"
as "main").
```

The type resolution introduced in #3233 was relying on either (a) the
symbol identifier table; or (b) the type not being in a submodule.

Unfortunately, there was a 3rd case which was not covered: the library
had already been imported by its entry point, and the
`this.submoduleMap` had been initialized will all exported submodules.

Restore that code path, and improve the error message for undeclared
dependencies.



---

By submitting this pull request, I confirm that my contribution is made under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0