chore: move disclaimer logic from pacmak to rosetta (#3308)
We add disclaimers to the visible source when we are translating examples that do not compile.
Previously this logic was in `pacmak`, and the main reason this was the case was because that
was where we knew what language we were translating to (and could append the correct
comment token). `pacmak` shouldn't really know/change snippets, so this PR refactors that logic
into `rosetta` instead.

---

By submitting this pull request, I confirm that my contribution is made under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0