chore(rosetta): loose option not exposed to command-line for extract (#3304)
In `ExtractOptions`, we specify an option called `loose`. We should also have a command-line
argument `--loose` to reflect that, which is what I'm adding here.

The `loose` flag is already tested; so I don't think we need to add any tests here.

---

By submitting this pull request, I confirm that my contribution is made under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0