feat(go): add UnsafeCast function
The `UnsafeCast` function can be used to forcefully convert from one
type of jsii proxy value (including `interface{}`) to another jsii
interface (i.e: a class or interface instance).

If performs a "clean" cast when possible, and creates a new, aliased
proxy otherwise. It is the user's responsibility to ensure they are
passing the correct arguments into the function.