feat(python): support inheritance-based interface implementation
The code generator used to represent jsii interfaces as python
`Protocols`, however this mechanism is normally intended to allow
structural typing in python. Python `Protocol` types are declared using
a custom metaclass, which makes them tricky to inherit from without
running into metaclass conflicts (especially in cases where multiple
inheritance involving a `Protocol` and a non-`Protocol` base class).

The jsii runtime does not perform structural typing, and hence
interfaces are better modelled as abstract base classes that only
declare abstract members.

This, together with a minor modification to the `@jsii.interface`
decorator, allows interfaces to be "implemented" by simply adding them
to the implementor's inheritance chain, as is "natural" to do in Python
in such cases.

The "legacy" `@jsii.implements` approach to implementing interfaces is
no longer recommended, however it is still supported as it is necessary
when dealing with libraries generated by older versions of
`jsii-pacmak`.