feat(go): Use type registry to find the correct target type for JSII Proxy (#3354)
With this PR, the type that is instantiated as JSII Proxy is looked up from the type registry using the TypeFQDN in the ref. If that type is registered in the type registry AND that type is assignable to the targeted value type, than that type is used to instantiate the JSII proxy. 

Fixes #3353 and relates to #2819.