chore: deprecate Python 3.6 support (#3364)
Python 3.6 has reached end-of-life at the end of December 2021, and new
`pip` releases no longer support it.

Check for python versions at load of the `jsii` runtime for Python, and
emit a warning if Python < 3.7 is detected. Additionally, change the
`pip` build-tool requirement to be `pip ~= 22.0` if Python >= 3.7 is
used.



---

By submitting this pull request, I confirm that my contribution is made under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0