fix: remove the "comments rewriter" (#3368)
The comments rewriter was used to prefix documentation comments with the
stability level of the associated API (e.g: `(experimental)` or
`(deprecated)`). This was introduced to palliate insufficient IDE
support for `@deprecated` and the mixing of `@stable` and
`@experimental` APIs in CDK v1.

Since then, the TypeScript language server has added support for
`@deprecated` IDE features, and CDK v2 moved all `@experimental` APIs to
separate packages, removing the risk for people to inadvertently take
experimental dependencies.

In profiling, this feature alone was attributed between 50% and 60% of
the total time spent compiling `aws-cdk-lib`. The impact is likely
comparable on other large construct libraries.

The benefits from this feature are not sufficient to justify the cost to
make it happen. Instead of trying to optimize it (which could prove
impossible), decision was made to simply drop it.



---

By submitting this pull request, I confirm that my contribution is made under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0