fix(jsii): excessive overrides declarations registered
Members inherited from parent base classes, which also implemented a
locally implemented interface member, would get tagged with the
`overrides` marker for the interface, which was incorrect, since the
declaring type did not implement the interface directly.

This is solved by passing the inherited members by-copy to the
validator, so side effects of it aren't visible to the final assembly.