fix(jsii): excessive overrides declarations registered (#3375)
Members inherited from parent base classes, which also implemented a
locally implemented interface member, would get tagged with the
`overrides` marker for the interface, which was incorrect, since the
declaring type did not implement the interface directly.

This is solved by passing the inherited members by-copy to the
validator, so side effects of it aren't visible to the final assembly.



---

By submitting this pull request, I confirm that my contribution is made under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0