fix(go): lots of documentation is missing from generated code
Most of the documentation text was incorrectly emitted on implementation
functions, which are not visible in documentation (as the structs that
are implementing the interfaces are not exported).

This moves the documentation text to the interface declarations instead,
so that they are visible in user-facing documentation.

Additionally, enum member documentation was simply not emitted, which is
addressed by this PR.

Setter documentations now only carry stability markers for the property
in order to avoid causing confusion by issuing the same documentation as
for the getter (which may be inappropriate).

This PR also adds transliterated example code to the generated
documentation, as the initial version of Rosetta support for Go is now
available.