fix(jsii): fail compilation when two or more enum members have same value
If two enum members have the same value, only the first one will
be retained.

This is a bit of an issue as we are renaming enum members: the named
version will never appear in the assembly, and so not work over jsii.

What's worse, if we deprecate-and-strip the original one, neither
of the enum members will appear.

Addressing the issue by failing the compilation by adding validation for
enum values

Fixes #2782.