feat: drop internal functions from Deprecation Error stack trace
JSII can throw Deprecation Error and fail compilation when a deprecated property,
enum or method is being used when `JSII_DEPRECATED` environment variable is set
to `fail`. The deprecation error is throw from `.warnings.jsii.js` generated file.
When nested prop of enum is deprecated, the stack trace will contain lot of methods
from `.warnings.jsii.js` and since the file is minimized, tools like Jest which try to be
helpful by printing the line where the error is thrown shows a wall of text without much useful
information.

With this change, the Deprecation error is caught and rethrown from top level class which
call the validator to provide a clean stack trace. This approch is taken instead of passing the
call site to deprecation handler as we can't get call site for setters and  getters that can be
used by `Error.captureStack`