perf: cache package.json and .jsiirc file lookup
JSII loads `package.json` and `.jsiirc` file multiple times in the compilation
process. These files are loaded from disk everytime and are expensive operation.
Updated the code to cache the file in memort by memoizing the load file calls.

With this change the compilation time for `aws-cdk-lib` module came from 133 seccond
to 101 second