chore: drop support for Python 3.6 (EOL as of 2021-12-23)
Python 3.6 has been end-of-life since late december 2021, and an
increasing amount of our dependencies are dropping support for it as a
consequence.

We introduced a warning to users that are still using Python 3.6 with
release `v1.55.1` around March 16, 2022.

We are now effectively dropping support for Python 3.6, and introducing
testing for Python 3.10 releases.

Related: https://github.com/aws/jsii/commit/919f735778fe57047fba45f66a821a4abbe28afc

BREAKING CHANGE: `jsii-pacmak` and the libraries it generates for Python
targets now require a minimum Python version of 3.7, instead of 3.6
previously. We recommend users upgrade to the latest supported Python
release (Python 3.10).