chore: cache the results of `findPackageInfo`
Profiling showed that `findPackageInfo` was one of the top contributors
to compiler duration, as it was reading the same (large) files over and
over again. It also incurred additional GC churn (as those duplicated
large data blobs would need to be recclaimed).

This change adds a memoization map around this call to avoid re-reading
those files during a single compiler pass, which reduces GC overhead as
well as file IO overhead.