feat(go): add tag to denote required/optional properties
Adds a `required` or `optional` tag on struct fields to denote whether
they are required or optional (obviously). This is a cheap-ass way to
provide documentation visibility to users.

This isn't perfect as ideally, one would want compile-time enforcement
of these, however this isn't currently posisble with Go (we explored
several options, including via the use of generics, but it did not end
up with any particularly improved solution).

Fixes #2536