fix(go): object type not recognized (#3485)
When a struct pointer was passed in an `interface{}` position, the
converter would trip and register a new opaque instance of `Object`
instead of correctly serializing a struct.

This is because hte pointer is double-indirected (the `interface{}` is a
reference to the `&struct`, which is a reference to the `struct`). This
change adds a condition to detect such cases (`interface{}` abstracts
another pointer), and properly serialize that.

Fixes #2880



---

By submitting this pull request, I confirm that my contribution is made under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0