chore: automated node runtime deprecation warnings
This changes how the node release support information is modeled so that
the end-of-life dates are configured, so that `@jsii/check-node` can
automatically start warning a month ahead of a runtime's EOL date, so we
can drop support right on EOL day if needed wihtout taking users by
surprise.

This also includes tweaks of the messages that get printed out to STDERR
when an unsupported/untested/deprecated/end-of-life version is being
used, to increase clarity of provide more information about migration
options.