fix(pacmak): missing python artifacts when using relative -o path
When the `-o`/`--outdir` option is passed to `jsii-pacmak` with a
relative path, that path was not resolved up-front by the CLI, which
resulted in the Python code-generator not outputting the wheel files in
the expected directory: instead of being consistently treated as
relative to the current working directory, the Python generator
interpreted it relative to it's temporary working directory.

This addresses the issue by resolving relative paths to absolute paths,
using `process.cwd()` as the base directory, before passing the value
down to the business logic.

Fixes aws/aws-cdk#20041